+++ src/Rask.Templates/server/Features/Auth/RegisterPage.cs
@@ -1,7 +1,5 @@
 using Company.RaskServer.Features.Shared;
 using Microsoft.AspNetCore.Authorization;
-using Rask.Core.Authentication;
-using Rask.Core.Routing;
 using Rask.Wire;
 
 namespace Company.RaskServer.Features.Auth;