+++ src/Rask.Templates/wasm-hosted/Client/App.cs
@@ -1,7 +1,3 @@
-using Rask.Core.Live;
-using Rask.Core.Routing;
-using Rask;
-
 namespace Company.RaskServer.Client;
 
 public sealed partial class App : Component